// src/app/models/cv.model.ts
export interface Cv {
    id: number;           // Unique identifier for the CV
    name: string;         // Name of the person
    job: string;          // Job title
    description: string;  // Brief description about the person
    avatar: string;       // URL or path to the person's avatar image
  }
  